rfac: code in toolbar & medialayer manager
