--- util/GlassLayer.js
@@ -1,85 +1,0 @@
-import * as THREE from "three";
-
-class GlassLayer {
-    constructor(layer, renderer) {
-        this.layer = layer;
-        this.renderer = renderer;
-
-        this.glassObject = this.createGlassObject(this.layer);
-    }
-
-    get object() {
-        return this.glassObject;
-    }
-
-    createGlassObject({
-        width,
-        height,
-        transform: {
-            position: { x, y, z },
-        },
-    }) {
-        const planeGeometry = new THREE.PlaneGeometry(1, 1); // to scale
-        const planeMaterial = new THREE.MeshBasicMaterial({
-            transparent: true,
-            opacity: 0.5, // test
-        });
-        const plane = new THREE.Mesh(planeGeometry, planeMaterial);
-
-        plane.scale.set(2 * width, 2 * height, 1);
-        plane.position.set(x, y, z + 0.01);
-        return plane;
-    }
-
-    move({ x, y, z }) {
-        this.glassObject.position.x = x;
-        this.glassObject.position.y = y;
-        this.glassObject.position.z = z;
-        this.glassObject.position.needsUpdate = true;
-    }
-
-    updateDimensions({ width, height }) {
-        this.glassObject.scale.set(2 * width, 2 * height, 1);
-    }
-
-    /**
-     * Updates position and quaternion of glass layer when quad video layer is moved
-     */
-    updateOrientation(position, quaternion) {
-        // update position x, y, z
-        const { x, y, z } = position;
-        this.glassObject.position.x = x;
-        this.glassObject.position.y = y;
-        this.glassObject.position.z = z;
-
-        // update quaternion (3d heading and orientation)
-        this.glassObject.quaternion.copy(quaternion);
-    }
-
-    updateOnRender() {
-        this.updateDimensions(this.layer);
-    }
-
-    move() {
-        // const position = new THREE.Vector3();
-        // const quaternion = new THREE.Quaternion();
-        // this.glassObject.getWorldPosition(position);
-        // this.glassObject.getWorldQuaternion(quaternion);
-        // const { x, y, z } = position;
-        // this.layer.transform = new XRRigidTransform(
-        //     {
-        //         x,
-        //         y,
-        //         z,
-        //         w: 1.0,
-        //     },
-        //     quaternion
-        // );
-        this.updateOrientation(
-            this.layer.transform.position,
-            this.layer.transform.orientation
-        );
-    }
-}
-
-export default GlassLayer;
